+++ addon/index.js
@@ -1,2 +1,5 @@
 import UrlTemplates from "ember-data-url-templates/mixins/url-templates";
+import UrlTemplatesSerializerMixin from "ember-data-url-templates/mixins/url-templates-serializer";
+
+export { UrlTemplatesSerializerMixin as UrlTemplatesSerializer };
 export default UrlTemplates;