+++ tests/dummy/app/adapters/post.js
@@ -2,7 +2,12 @@ import ApplicationAdapter from './application';
 
 export default ApplicationAdapter.extend({
   urlTemplate: "{+host}{/namespace}/my-posts",
+  findRecordUrlTemplate: "{+host}{/namespace}/posts/{id}",
   queryRecordUrlTemplate: "{+host}{/namespace}/my-posts/{slug}",
   updateRecordUrlTemplate: "{+host}{/namespace}/my-posts/{slug}",
+
+  authorUrlTemplate: "{+host}{/namespace}/posts/{id}/author",
+  reactionsUrlTemplate: "{+host}{/namespace}/posts/{id}/reactions",
+
   namespace: 'api',
 });