+++ tests/dummy/app/serializers/post.js
@@ -1,0 +1,6 @@
+import DS from 'ember-data';
+
+import { UrlTemplatesSerializer } from "ember-data-url-templates";
+
+export default DS.JSONAPISerializer.extend(UrlTemplatesSerializer, {
+});