+++ tests/dummy/mirage/factories/post.js
@@ -1,6 +1,8 @@
-import { Factory, faker } from 'ember-cli-mirage';
+import { Factory, faker, association } from 'ember-cli-mirage';
 
 export default Factory.extend({
+  author: association(),
+
   slug() {
     return faker.helpers.slugify(this.title);
   },