+++ tests/dummy/mirage/models/post.js
@@ -1,5 +1,7 @@
-import { Model, hasMany } from 'ember-cli-mirage';
+import { Model, hasMany, belongsTo } from 'ember-cli-mirage';
 
 export default Model.extend({
   comments: hasMany(),
+  author: belongsTo(),
+  relatedPosts: hasMany('post'),
 });